refactor: move mcp_service to new mcp folder
