+++ apps/backend/src/rhesis/backend/app/routers/services.py
@@ -39,7 +39,7 @@ from rhesis.backend.app.services.generation import (
     generate_tests,
 )
 from rhesis.backend.app.services.github import read_repo_contents
-from rhesis.backend.app.services.mcp_service import (
+from rhesis.backend.app.services.mcp import (
     extract_mcp,
     handle_mcp_exception,
     query_mcp,
