+++ apps/backend/src/rhesis/backend/app/services/mcp/__init__.py
@@ -1,0 +1,17 @@
+"""MCP services for external integrations."""
+
+from rhesis.backend.app.services.mcp.service import (
+    extract_mcp,
+    handle_mcp_exception,
+    query_mcp,
+    run_mcp_authentication_test,
+    search_mcp,
+)
+
+__all__ = [
+    "extract_mcp",
+    "handle_mcp_exception",
+    "query_mcp",
+    "run_mcp_authentication_test",
+    "search_mcp",
+]
